[Fix #1226]  Waiting for pending writes before close
